requirements.txt
streamlit
playwright
google-generativeai
Pillow